src/ai-service/requirements.txt
# „Der Herr, unser Gott, lasse uns freundlich ansehen. Lass unsere Arbeit nicht vergeblich sein – ja, lass gelingen, was wir tun!" Psalm 90,17

# Core dependencies [CAM][MLB]
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
psutil==5.9.6  # System monitoring for health checks [ATV]

# Whisper transcription
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

# Anonymization
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2
de_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.7.0/de_core_news_lg-3.7.0-py3-none-any.whl

# AI Providers
openai==1.3.5
anthropic==0.5.0
google-generativeai==0.3.1

# Security & Logging [SP][ATV][DSC]
cryptography==41.0.5
python-jose==3.3.0
structlog==23.2.0
pytz==2023.3  # Timezone support for Swiss date formatting [SF]
frozendict==2.3.8  # Thread-safe dictionaries for metrics collection [ATV]

# Testing [KP100][ATV]
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0  # Better mocking for unit tests
httpx==0.25.1
grpc-testing==1.59.0  # gRPC testing utilities

# Code quality
black==23.10.1
isort==5.12.0
mypy==1.6.1
